feat(settingsitem): move all tab settings to settingsitem
