+++ app/components/Home/modals/SettingsItems/SettingsItem.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+
+type SettingsItemProps = {
+  label: string;
+  children: React.ReactNode;
+};
+
+const SettingsItem: React.FC<SettingsItemProps> = ({ label, children }) => {
+  return (
+    <div className="text-2xl flex flex-col gap-2 font-semibold">
+      <span>{label}</span>
+      <div>{children}</div>
+    </div>
+  );
+};
+
+export default SettingsItem;
