+++ app/components/Home/modals/ToonSettingsModal.tsx
@@ -3,6 +3,7 @@ import Modal from "../../Modal";
 import { StoredToonData } from "@/app/types";
 import { useToonContext } from "@/app/context/ToonContext";
 import { FaLock, FaUnlock, FaCog, FaTrash } from "react-icons/fa";
+import GardenSettingsItem from "./SettingsItems/GardenSettingsItem";
 
 type SettingsModalProps = {
   toon: StoredToonData | null;
@@ -19,14 +20,6 @@ const SettingsModal: React.FC<SettingsModalProps> = ({
 }) => {
   if (!isOpen || !toon || index == null) return null;
   const { toons, addToon, deleteToon } = useToonContext();
-  const [flowerType, setFlowerType] = useState<1 | 2 | 3>(() => {
-    return JSON.parse(localStorage.getItem("flowerType") || "1");
-  });
-
-  useEffect(() => {
-    localStorage.setItem("flowerType", JSON.stringify(flowerType));
-    window.dispatchEvent(new Event("flowerChange"));
-  }, [flowerType]);
 
   const toggleLock = (index: number) => {
     const toon = toons[index];
@@ -37,10 +30,6 @@ const SettingsModal: React.FC<SettingsModalProps> = ({
   const getLockedStatus = (index: number) => {
     return toons[index]?.locked ?? false;
   };
-
-  function handleFlowerDisplay(arg: 1 | 2 | 3): void {
-    setFlowerType(arg);
-  }
 
   return (
     <Modal isOpen={isOpen} onClose={onClose}>
@@ -72,62 +61,13 @@ const SettingsModal: React.FC<SettingsModalProps> = ({
             Global Settings
           </h3>
 
-          {/* gardening settings */}
+          {/* fish settings */}
           <div className="text-2xl flex items-center gap-2 font-semibold">
-            <span>Gardening Display</span>
+            <span>Fishing</span>
           </div>
-          <div>
-            <div className="flex items-center space-x-2">
-              <input
-                type="radio"
-                name="flowerType"
-                id="progressFlowers"
-                checked={flowerType === 1}
-                onChange={() => handleFlowerDisplay(1)}
-                className="w-5 h-5 cursor-pointer"
-              />
-              <label
-                className="text-lg cursor-pointer hover:text-blue-600"
-                htmlFor="progressFlowers"
-              >
-                Only flowers that grant progress
-              </label>
-            </div>
 
-            <div className="flex items-center space-x-2">
-              <input
-                type="radio"
-                name="flowerType"
-                id="plantableFlowers"
-                checked={flowerType === 2}
-                onChange={() => handleFlowerDisplay(2)}
-                className="w-5 h-5 cursor-pointer"
-              />
-              <label
-                className="text-lg cursor-pointer hover:text-blue-600"
-                htmlFor="plantableFlowers"
-              >
-                All plantable flowers
-              </label>
-            </div>
-
-            <div className="flex items-center space-x-2">
-              <input
-                type="radio"
-                name="flowerType"
-                id="allFlowers"
-                checked={flowerType === 3}
-                onChange={() => handleFlowerDisplay(3)}
-                className="w-5 h-5 cursor-pointer"
-              />
-              <label
-                className="text-lg cursor-pointer hover:text-blue-600"
-                htmlFor="allFlowers"
-              >
-                All flowers
-              </label>
-            </div>
-          </div>
+          {/* gardening settings */}
+          <GardenSettingsItem />
         </div>
 
         {/* deletion */}
